DetailsTableViewCell: set the images and the titles.
DetailsViewController:
set nib view
set the data and pass to the titles and the image.

MainViewController:
set the data that will request by the func in RequestManager.
set TableView.

ObjectManager:
created a structs for json request.

RequestManager:
get data from json file and pass to DetailsVC and MainVC
download session of images. and pass to DetailsVC and MainVC